requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
pandas>=2.0.0
numpy>=1.24.0
seaborn>=0.12.0
openai>=1.0.0
langchain>=0.1.0
langchainhub>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.5
langchain_experimental>=0.0.40
langchain_cohere>=0.0.3
tabulate>=0.9.0
black>=23.0.0
isort>=5.12.0

